requirements.txt
annotated-types==0.7.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.3
cycler==0.12.1
cymem==2.0.11
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
fonttools==4.60.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
murmurhash==1.0.13
nltk==3.9.2
numpy==2.3.4
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
preshed==3.0.10
pydantic==2.12.2
pydantic_core==2.41.4
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
rake-nltk==1.0.6
regex==2025.9.18
requests==2.32.5
rich==14.2.0
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
textblob==0.19.0
thinc==8.3.6
threadpoolctl==3.6.0
tqdm==4.67.1
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
Unidecode==1.4.0
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.3
